Add files via upload
The first Jupyter notebook contains code to extract lyrics corresponding to a particular singer (see EP.py in the main directory or the subsequent .py files within the folders corresponding to singers). The second Jupyter notebook corresponds to the code in 'lyrics_scraping_project.py' file to label the text and train a text classification algorithm.